requirements.txt

openai
python-dotenv
PyMuPDF
langchain
langchain-community
chromadb
